backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
django-environ==0.11.2
requests==2.31.0
Pillow==10.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-celery-beat==2.5.0
pytz==2024.1
mysqlclient==2.2.4
cryptography==42.0.5
PyJWT==2.8.0
